plash_deploy/requirements.txt
python-fasthtml
MonsterUI
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-vector-stores-qdrant
llama-index-postprocessor-cohere-rerank
qdrant-client
openai
cohere
fastapi

